doc(README): fix flang-new testing instructions
In order to test command_line_t, the test suite must run fpm via
execute_command_line.  With flang-new, this requires ensuring that
the fpm flags are passed along to the so-invoked fpm commands.
This commit updateds the README to direct uses to set FPM_FC and
FPM_FFLAGS to ensure that the fpm commands launched by the test
suite receive the necessary flags.